Add network statistics gathering to the LocalJmxMonitor
Added new threads to capture the number of inboud and outbound bytes for
an interface. (NetworkBytesMonitor and subclasses NetworkBytesInMonitor
and NetworkBytesOutMonitor) The /proc/net/dev code was shamelessly
stolen from Michal's SimpleNetworkTrafficCollector SmartFrog class.

Add references to the new threads in LocalJmxMonitor, and only enable
them when the interface name field is set.

Add interface name property to StartJVMMonitorStage and set the field in
LocalJmxMonitor when the property is specified.

Plot the network statistics in LocalSystemMonitorChart