Changes to AbstractDistStage and StartJVMMonitorStage
I noticed that StartJVMMonitorStage was not initializing the slave index
correctly. I also noticed that both StartJVMMonitorStage and
JPRofilerStage were overriding initOnMaster to initialize the configName
variable, so I added the field and initialization to AbstractDistStage.

Added the ability to set the frequency that JVM statistics are collected
in the StartJVMMonitorStage. The statistics are collected once a second,
minute, or hour.