Addressed review comments
StartJVMMonitorStage now take frequency and time unit parameters
LocalJmxMonitor stores the frequency as well as the time unit
ClusterTimeSeriesReport takes the frequency and time unit in the
constructor, and generates a series based on it.
LocalSystemMonitorChart gets the frequency and time unit from
LocalJmxMonitor and constructs ClusterTimeSeriesReport objects using
these parameters. It also uses the correct labels in reports and charts.