Minor changes to Map/Reduce and Random Data stages
RandomDataStage - allow the memory overhead to be set, so that it can be
used with different products and versions. Also allow specifying the
total data size, and letting the stage calculate the number of values to
put into the cache.

MapReduceStage - add a parameter to print the result map and object