Add support for MapReduceTask combiners
Infinispan 5.2 added the ability to specify a Reducer that would be used
to reduce the values on the local node before the global reduce phase is
executed. This commit adds support for specifying this option in the
MapReduceStage parameters and implements support for them in the
MapReduceCapable interface and InfinispanMapReduce class.