Update code to detect cache rehashing in Infinispan Server Service
This change checks the pendingViewAsString attribute of the RpcManager
component associated with each cache using JMX. If pendingViewAsString
equals "null", no topology change is in progress. This commit keeps
track of a topology change happening on each cache, as well as across
all caches.

This change also checks the stateTransferInProgress attributes of the
StateTransferManager component associated with each cache using JMX. If
stateTransferInProgress is true, then a rehash is in progress. This
commit keeps track of a topology change happening on each cache, as well
as across all caches.