Allow InfinispanServerService to use an existing configuration file
Updated the init method to look for the specified configuration file in
the specified server's configuration directory. If it is found there,
then it is copied to 'radargun-${slaveIndex}.xml', and the file variable
is updated to the full path to the configuration file. This allows the
configuration dumping to work correctly.