Fix issue 314
The previous strategy of checking for processes based on the shell
process was not working, because the new java processes do not have the
expected parent PID. Instead tail_log takes a new argument that is a
search string for the 'ps -ef' command. If the read times out and the
process doesn't exist, then the shell process exits. Also added some
additional exit strings to master.sh.

https://github.com/radargun/radargun/issues/314