Add plugins for snapshot builds and use parent POMs for versions
Adding the Infinispan or JDG parent POM in the dependencyManagement
section means that the version is pulled from the parent POM directly.

Added snapshot build plugins for JDG and Infinispan. The Infinispan and
JDG snapshot build plugins are in separate profiles.

Move the properties and profiles from the RadarGun top-level POM to the
parent POM, so that these properties can be referenced.